agents/retriever_agent/requirements-retriever.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # or faiss-gpu if CUDA is available
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
python-dotenv>=1.0.0
